backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg[binary]==3.2.3
alembic==1.13.3
pydantic-settings==2.6.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
apscheduler==3.10.4
elasticsearch==8.15.1
python-dateutil==2.9.0.post0
